일부 수정
